Switched state management to Redux
