+++ src/components/Login.jsx
@@ -4,6 +4,13 @@ import { Button } from 'react-bootstrap'
 export default function Login() {
 
   return (
-    <Button type="link" href={process.env.REACT_APP_LOGIN_LINK} variant="black" className="login">Login to Spotify</Button>    
+    <Button 
+      type="link" 
+      href={process.env.REACT_APP_LOGIN_LINK} 
+      variant="black" 
+      className="login"
+    >
+      Login to Spotify
+    </Button>    
   )
 }
