+++ src/redux/types.js
@@ -1,0 +1,2 @@
+export const GET_LOGGED_IN = "GET_LOGGED_IN"; 
+export const FETCH_NOW_PLAYING = "FETCH_NOW_PLAYING";
